Move strtobool to try to resolve circular dependency
