+++ buildpack/telemetry/datadog.py
@@ -22,6 +22,7 @@ from buildpack import util
 from buildpack.core import runtime
 from buildpack.infrastructure import database
 from lib.m2ee.version import MXVersion
+from lib.m2ee.util import strtobool
 
 NAMESPACE = "datadog"
 TRACE_AGENT_DEPENDENCY = f"{NAMESPACE}.trace-agent"
@@ -74,28 +75,26 @@ def is_enabled():
 
 # Toggles Datadog APM
 def _is_tracing_enabled():
-    return util.strtobool(os.environ.get("DD_TRACE_ENABLED", "false"))
+    return strtobool(os.environ.get("DD_TRACE_ENABLED", "false"))
 
 
 # Toggles logs redaction (email addresses are replaced by a generic string)
 def _is_logs_redaction_enabled():
-    return util.strtobool(os.environ.get("DATADOG_LOGS_REDACTION", "true"))
+    return strtobool(os.environ.get("DATADOG_LOGS_REDACTION", "true"))
 
 
 # Toggles database rare / count metrics which are collected by Telegraf
 # By default, they are not compatible with the Datadog Postgres integration
 # due to Telegraf limitations
 def is_database_rate_count_metrics_enabled():
-    return util.strtobool(
-        os.environ.get("DATADOG_DATABASE_RATE_COUNT_METRICS", "false")
-    )
+    return strtobool(os.environ.get("DATADOG_DATABASE_RATE_COUNT_METRICS", "false"))
 
 
 # Toggles the database diskstorage metrics
 # It is basically a fixed value based on an environment variable
 def is_database_diskstorage_metric_enabled():
     return (
-        util.strtobool(os.environ.get("DATADOG_DATABASE_DISKSTORAGE_METRIC", "true"))
+        strtobool(os.environ.get("DATADOG_DATABASE_DISKSTORAGE_METRIC", "true"))
         and os.environ.get("DATABASE_DISKSTORAGE") is not None
     )
 
@@ -103,7 +102,7 @@ def is_database_diskstorage_metric_enabled():
 # Toggles the system checks. Note that these may not mean anything as they
 # might show the host metrics instead of the container metrics.
 def _is_checks_enabled():
-    return util.strtobool(os.environ.get("DD_ENABLE_CHECKS", "false"))
+    return strtobool(os.environ.get("DD_ENABLE_CHECKS", "false"))
 
 
 # Toggles Datadog profiling. Can only enabled when using AdoptOpenJDK and
@@ -111,7 +110,7 @@ def _is_checks_enabled():
 def _is_profiling_enabled(runtime_version):
     if runtime_version < MXVersion("7.23.1") or not _is_tracing_enabled():
         return False
-    return util.strtobool(os.environ.get("DD_PROFILING_ENABLED", "false"))
+    return strtobool(os.environ.get("DD_PROFILING_ENABLED", "false"))
 
 
 def _is_installed():
