+++ buildpack/telemetry/metrics.py
@@ -16,6 +16,7 @@ from buildpack.core import runtime
 from buildpack.infrastructure import database
 from lib.m2ee import munin
 from lib.m2ee.version import MXVersion
+from lib.m2ee.util import strtobool
 
 from . import datadog, appdynamics, dynatrace
 
@@ -139,9 +140,7 @@ def get_micrometer_metrics_url():
     micrometer metrics only. Runtime version 9.7 and above is required.
 
     """
-    use_trends_forwarder = util.strtobool(
-        os.getenv("USE_TRENDS_FORWARDER", default="true")
-    )
+    use_trends_forwarder = strtobool(os.getenv("USE_TRENDS_FORWARDER", default="true"))
 
     trends_forwarder_url = os.getenv("TRENDS_FORWARDER_URL", default="")
 
@@ -160,9 +159,7 @@ def _micrometer_runtime_requirement(runtime_version):
     # TODO: DISABLE_MICROMETER_METRICS is a temporary flag to disable metrics
     # collection via micrometer till we are ready to do the switchover
     # from admin port metrics to micrometer based metrics
-    disable_micrometer = util.strtobool(
-        os.getenv("DISABLE_MICROMETER_METRICS", "false")
-    )
+    disable_micrometer = strtobool(os.getenv("DISABLE_MICROMETER_METRICS", "false"))
 
     runtime_version_supported = runtime_version >= MXVERSION_MICROMETER
 
@@ -209,7 +206,7 @@ def bypass_loggregator():
     # Throws a useful message if you put in a nonsensical value.
     # Necessary since we store these in cloud portal as strings.
     try:
-        bypass = util.strtobool(env_var)
+        bypass = strtobool(env_var)
     except ValueError:
         logging.warning(
             "Bypass loggregator has a nonsensical value: %s. "
