+++ buildpack/telemetry/telegraf.py
@@ -15,6 +15,7 @@ import subprocess
 from buildpack import util
 from buildpack.core import runtime
 from buildpack.infrastructure import database
+from lib.m2ee.util import strtobool
 from jinja2 import Template
 
 from . import datadog, metrics, mx_java_agent, appdynamics, dynatrace
@@ -90,7 +91,7 @@ def include_db_metrics():
         # For customers who have Datadog or AppDynamics or APPMETRICS_TARGET enabled,
         # we always include the database metrics. They can opt out
         # using the APPMETRICS_INCLUDE_DB flag
-        result = util.strtobool(os.getenv("APPMETRICS_INCLUDE_DB", "true"))
+        result = strtobool(os.getenv("APPMETRICS_INCLUDE_DB", "true"))
 
     return result
 
@@ -262,7 +263,7 @@ def update_config(m2ee, app_name):
         dynatrace_enabled=dynatrace.is_telegraf_enabled(),
         dynatrace_config=_get_dynatrace_config(app_name),
         telegraf_debug_enabled=os.getenv("TELEGRAF_DEBUG_ENABLED", "false"),
-        telegraf_fileout_enabled=util.strtobool(
+        telegraf_fileout_enabled=strtobool(
             os.getenv("TELEGRAF_FILEOUT_ENABLED", "false")
         ),
     )
