+++ buildpack/util.py
@@ -730,22 +730,6 @@ def upsert_logging_config(m2ee, value):
     _upsert_m2ee_config_section(m2ee, "logging", [value], overwrite=False, append=True)
 
 
-# Replacement for deprecated distutils.util.strtobool function
-# And by replacement I mean a direct copy
-def strtobool(val):
-    """Convert a string representation of truth to true (1) or false (0).
-    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
-    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
-    'val' is anything else.
-    """
-    val = val.lower()
-    if val in ("y", "yes", "t", "true", "on", "1"):
-        return 1
-    if val in ("n", "no", "f", "false", "off", "0"):
-        return 0
-    raise ValueError(f"invalid truth value {val}")
-
-
 # Script entry point. Only meant to be used in development
 if __name__ == "__main__":
     import sys
